Changed directory structure, fixed pipeline and code
