+++ .gitignore
@@ -3,6 +3,11 @@
 ##
 ## Get latest from https://github.com/github/gitignore/blob/master/VisualStudio.gitignore
 
+# AML-specific files
+.azureml
+run.json
+model.json
+
 # User-specific files
 *.suo
 *.user
