--- deploymentConfig.yml
@@ -1,5 +1,0 @@
----
-containerResourceRequirements:
-  cpu: 1
-  memoryInGB: 1
-computeType: ACI
