--- deploymentconfig.json
@@ -1,7 +1,0 @@
-{
-   "containerResourceRequirements": {
-      "cpu": 1,
-      "memoryInGB": 1
-   },
-   "computeType": "ACI"
-}
