--- inferenceconfig.json
@@ -1,10 +1,0 @@
-{
-   "entryScript": "score.py",
-   "runtime": "python",
-   "condaFile": "myenv.yml",
-   "extraDockerfileSteps": null,
-   "sourceDirectory": null,
-   "enableGpu": false,
-   "baseImage": null,
-   "baseImageRegistry": null
-}
