+++ models/diabetes/aksDeploymentConfig.yml
@@ -1,16 +1,19 @@
 computeType: AKS
+
 autoScaler:
     autoscaleEnabled: True
     minReplicas: 1
     maxReplicas: 3
     refreshPeriodInSeconds: 10
     targetUtilization: 70
-authEnabled: True
+
 containerResourceRequirements:
     cpu: 1
     memoryInGB: 0.5
-appInsightsEnabled: False
-scoringTimeoutMs: 1000
-maxConcurrentRequestsPerContainer: 2
-maxQueueWaitMs: 1000
+
 sslEnabled: False
+authEnabled: True
+appInsightsEnabled: False
+maxQueueWaitMs: 1000
+scoringTimeoutMs: 1000
+maxConcurrentRequestsPerContainer: 1
