+++ models/diabetes/deploymentConfig.yml
@@ -1,0 +1,4 @@
+containerResourceRequirements:
+  cpu: 1
+  memoryInGB: 1
+computeType: ACI
