+++ pipelines/diabetes-train-and-deploy.yml
@@ -1,0 +1,74 @@
+variables:
+  ml-ws-connection: 'azmldemows' # Workspace Service Connection name
+  ml-ws: 'aml-demo' # AML Workspace name
+  ml-rg: 'aml-demo' # AML resource Group name
+  ml-ct: 'cpu-cluster-1' # AML Compute cluster name
+  ml-path: 'models/diabetes' # Model directory path in repo
+  ml-exp: 'exp-test' # Experiment name
+  ml-model-name: 'diabetes-model' # Model name
+
+trigger:
+- master
+
+pool:
+  vmImage: 'Ubuntu-16.04'
+
+steps:
+- task: AzureCLI@1
+  displayName: 'Install the AML CLI'
+  inputs:
+    azureSubscription: $(ml-ws-connection)
+    scriptLocation: inlineScript
+    inlineScript: 'az extension add -n azure-cli-ml'
+
+- task: AzureCLI@1
+  displayName: 'Attach folder to workspace'
+  inputs:
+    azureSubscription: $(ml-ws-connection)
+    workingDirectory: $(ml-path)
+    scriptLocation: inlineScript
+    inlineScript: 'az ml folder attach -w $(ml-ws) -g $(ml-rg)'
+
+- task: AzureCLI@1
+  displayName: 'Create compute for training'
+  inputs:
+    azureSubscription: $(ml-ws-connection)
+    workingDirectory: $(ml-path)
+    scriptLocation: inlineScript
+    inlineScript: 'az ml computetarget create amlcompute -n $(ml-ct) --vm-size STANDARD_D2_V2 --max-nodes 1'
+  
+- task: AzureCLI@1
+  displayName: 'Submit script run'
+  inputs:
+    azureSubscription: $(ml-ws-connection)
+    workingDirectory: $(ml-path)
+    scriptLocation: inlineScript
+    inlineScript: 'az ml run submit-script -c sklearn --ct $(ml-ct) -e $(ml-exp) -t run.json train.py'
+
+# Add potential approval step before registration of model
+
+- task: AzureCLI@1
+  displayName: 'Register model'
+  inputs:
+    azureSubscription: $(ml-ws-connection)
+    workingDirectory: $(ml-path)
+    scriptLocation: inlineScript
+    inlineScript: 'az ml model register -n $(ml-model-name) -f run.json --asset-path outputs/ridge_0.95.pkl -t model.json'
+
+- task: AzureCLI@1
+  displayName: 'Deploy model to ACI for QA'
+  inputs:
+    azureSubscription: $(ml-ws-connection)
+    workingDirectory: $(ml-path)
+    scriptLocation: inlineScript
+    inlineScript: 'az ml model deploy -n diabetes-qa-aci -f model.json --ic inferenceConfig.yml --dc deploymentConfig.yml --overwrite'
+
+# Add potential automated tests
+
+#- task: AzureCLI@1
+#  displayName: 'Delete deployed service'
+#  inputs:
+#    azureSubscription: $(ml-ws-connection)
+#    workingDirectory: $(ml-path)
+#    scriptLocation: inlineScript
+#    inlineScript: 'az ml service delete -n acicicd'
